Dev: rename to CctpTransferStarter
